+++ database/migrations/testing/2026_07_14_000000_create_test_applied_template_table.php
@@ -11,7 +11,7 @@
  *
  * @author     Jacob Martella <me@jacobmartella.com>
  *
- * @since      1.1.0
+ * @since      1.6.0
  */
 
 declare( strict_types=1 );